[ADD]: Preloader parent class
