+++ src/phaser/scenes/multiplay/Preloader.js
@@ -1,6 +1,6 @@
-import SinglePreloader from "../singleplay/Preloader";
+import Preloader from "../parent/preloader";
 
-export default class MultiPreloader extends SinglePreloader {
+export default class MultiPreloader extends Preloader {
   completeLoading() {
     this.scene.start("multi");
   }
