+++ src/phaser/scenes/singleplay/Stage1.js
@@ -1,4 +1,4 @@
-import Stage from "./parent";
+import Stage from "../parent/stage";
 import { Level1 } from "../../../constants/enemyList";
 
 export default class Stage1 extends Stage {
